Refinement of initialization, and we have developed some rules:
1. Only import the subsystem, not the individual plugin. This
will ensure that subsystem initialization happens correctly.

2. Set a model via hub.set_model(canonical_path, **kwargs).
The kwargs will be sent to the __init__ function in init.py
for the subsystem. You cannot set a model for an individual
plugin, just the subsystem.

3. Hubs are getting automatically mapped into every plugin.

4. Models are not getting automatically mapped -- they go to
the init.py __init__() function.

5. init.py is a special module. You can't "see" it as a plugin.
It is hidden, and is ad-hoc loaded to process the __init__()
function, directly by the PluginDirectory.