I encountered a pretty horrendous problem where my distfile-spider was showing messages like this in traceback:
AttributeError: module 'dyne.org.funtoo.metatools.merge.fastpull' has no attribute 'expand_uris'

The problem was that this function did exist in the plugin! So
this message was 'wrong'. What was going on?

I worked around it by putting asserts in the main thread to ensure
that these functions existed in the plugin. This worked around the
issue.

I then figured out that the issue was lack of thread-safety in the
plugin loader. I thought I had that issue resolved. I have added
a factory for PluginSubsystems to ensure that all dyne imports
return the same instance of PluginSubsystem. I also added an
access lock for loading plugins so those are sane across threads.

This seems to have resolved the issue.